Moved Feed to main components folder
