+++ client/src/Components/Feed.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import Spinner from "../Spinner";
+import Spinner from "./Spinner";
 import FeedTweet from "./FeedTweet";
 
 const Feed = ({ feedUrl }) => {
